Recent Fragment and Streaming Improve
Instead of directly using TorrentStream library, I moved to TorrentStream Server. With we should be able to provide a better handle of buffering and fix writing/reading permission on file.

Made the recent fragment in the same style as torrentlist fragment for now. Torrents are added to a RecentlyPlayed object which holds a torrents string property. This property with the help of the helpers provided can then be converted to JSONArray. You can add an item to recentplaylist with the helper function addTorrentToRecenPlaylist

Also added some strings to translate and updated gradlew to the last 4.1